Change cart ajax & json response, ajax done

+++ app1/carts/views.py
@@ -10,7 +10,6 @@ from goods.models import Products
 def cart_add(request):
 
     product_id = request.POST.get("product_id")
-    print("Product ID:", product_id)
     product = Products.objects.get(id=product_id)
 
     if request.user.is_authenticated:
@@ -39,9 +38,28 @@ def cart_add(request):
 def cart_change(request, product_slug):
     ...
 
-def cart_remove(request, cart_id):
+
+def cart_remove(request):
+    cart_id = request.POST.get("cart_id")
 
     cart = Cart.objects.get(id=cart_id)
-    cart.delete()
 
-    return redirect(request.META['HTTP_REFERER'])
+    quantity = cart.quantity
+
+    if cart.quantity > 1:
+        cart.quantity -= 1
+        cart.save()
+    else:
+        cart.delete()
+
+    user_cart = get_user_carts(request)
+    cart_items_html = render_to_string(
+        "carts/includes/included_cart.html", {"carts": user_cart}, request=request)
+
+    response_data = {
+        "message": "Товар уменьшен в корзине" if cart.quantity > 0 else "Товар удален из корзины",
+        "cart_items_html": cart_items_html,
+        "quantity_deleted": quantity,
+    }
+
+    return JsonResponse(response_data)
